References Oppevara/oppevara#42, partial implementation.
The simple parts are mostly done. This includes a fully functional
display of the created content, including building of a resulting table.
The code to check the responses and give feedback is still missing (will
be added once the Editor part is fully functional). The editor part is
missing some styling to display rows and columns next to each other. It
is also missing the last dynamic part, where the table is shown,
allowind creator to mark correct choices for the checkbox/radio button
elements. Some other issues could still be present.